Cor nos polignos randomica
Esse commit adiciona cores randomicas quando um poligno é criado.

+++ n3/objetografico.h
@@ -56,7 +56,7 @@ class ObjetoGrafico
 		 * R
 		 * Nivel de vermelho do objeto
 		 */
-		int r, 
+		float r, 
 		/**
 		 * G
 		 * Nivel de verde do objeto
